linting and refactors
