+++ src/mud_examples/inv.py
@@ -1,15 +1,28 @@
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
-#!/usr/env/bin python
 
 import argparse
 import logging
 import os
 import sys
 
-import numpy as np
-from scipy.stats import norm, uniform # The standard Normal distribution
-from scipy.stats import gaussian_kde as kde # A standard kernel density estimator
 import matplotlib.pyplot as plt
+import numpy as np
+from mud import __version__ as __mud_version__
+from scipy.stats import \
+    gaussian_kde as kde  # A standard kernel density estimator
+from scipy.stats import norm, uniform  # The standard Normal distribution
+
+from mud_examples import __version__
+
+_logger = logging.getLogger(__name__) # TODO: make use of this instead of print
+
+from matplotlib import cm
+from mud.funs import map_sol, mud_sol
+from mud.norm import full_functional, norm_data, norm_input, norm_predicted
+
+from mud_examples.parsers import parse_args
+from mud_examples.utils import check_dir, make_2d_unit_mesh
 
 plt.rcParams['figure.figsize'] = 10,10
 plt.rcParams['font.size'] = 24
@@ -17,17 +30,6 @@ plt.rcParams['font.size'] = 24
 __author__ = "Mathematical Michael"
 __copyright__ = "Mathematical Michael"
 __license__ = "mit"
-from mud_examples import __version__
-from mud import __version__ as __mud_version__
-
-_logger = logging.getLogger(__name__) # TODO: make use of this instead of print
-
-from mud_examples.helpers import make_2d_unit_mesh, check_dir
-from mud.norm import full_functional, norm_input, norm_data, norm_predicted
-from matplotlib import cm
-from mud.funs import mud_sol, map_sol
-
-from mud_examples.helpers import parse_args
 
 def setup_logging(loglevel):
     """Setup basic logging
