+++ src/mud_examples/models.py
@@ -1,4 +1,6 @@
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
+
 import numpy as np
 
 
@@ -9,13 +11,15 @@ def generate_decay_model(t, lam_true):
         initial_cond = 0.75
         rate = lam[:, 0].reshape(-1, 1)
         response = initial_cond * np.exp(np.outer(rate, -t))
+        # this allows support for simpler 1D plotting.
         if response.shape[0] == 1:
-            return response.ravel()  # this allows support for simpler 1D plotting.
+            return response.ravel()
         return response
     return model
 
 
-def generate_temporal_measurements(measurement_hertz=100, start_time=1, end_time=3):
+def generate_temporal_measurements(measurement_hertz=100,
+                                   start_time=1, end_time=3):
     num_measure = measurement_hertz * (end_time - start_time)
     return np.linspace(start_time, end_time, num_measure)
 
@@ -31,8 +35,8 @@ def generate_spatial_measurements(num_measure,
 
 def generate_rotation_map(qnum=10, orth=True):
     if orth:
-        return np.array([[np.sin(theta), np.cos(theta)] for theta in
-                         np.linspace(0, np.pi, qnum + 1)[0:-1]]).reshape(qnum, 2)
+        vec = np.linspace(0, np.pi, qnum + 1)[0:-1]
     else:
-        return np.array([[np.sin(theta), np.cos(theta)] for theta in
-                         np.linspace(0, np.pi, qnum)]).reshape(qnum, 2)
+        vec = np.linspace(0, np.pi, qnum)
+    matrix_entries = [[np.sin(theta), np.cos(theta)] for theta in vec]
+    return np.array(matrix_entries).reshape(qnum, 2)
