+++ src/mud_examples/utils.py
@@ -1,0 +1,62 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+import numpy as np
+import os
+import types
+import importlib
+
+
+def check_dir(directory):
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
+
+def make_2d_normal_mesh(N=50, window=1):
+    X = np.linspace(-window, window, N)
+    Y = np.linspace(-window, window, N)
+    X, Y = np.meshgrid(X, Y)
+    XX = np.vstack([X.ravel(), Y.ravel()]).T
+    return (X, Y, XX)
+
+
+def make_2d_unit_mesh(N=50, window=1):
+    X = np.linspace(0, window, N)
+    Y = np.linspace(0, window, N)
+    X, Y = np.meshgrid(X, Y)
+    XX = np.vstack([X.ravel(), Y.ravel()]).T
+    return (X, Y, XX)
+
+
+def fit_log_linear_regression(input_values, output_values):
+    x, y = np.log10(input_values), np.log10(output_values)
+    X, Y = np.vander(x, 2), np.array(y).reshape(-1, 1)
+    slope, intercept = (np.linalg.pinv(X) @ Y).ravel()
+    regression_line = 10**(slope * x + intercept)
+    return regression_line, slope
+
+
+class LazyLoader(types.ModuleType):
+    def __init__(self, module_name='utensor_cgen', submod_name=None):
+        self._module_name = '{}{}'.format(
+            module_name,
+            submod_name and '.{}'.format(submod_name) or ''
+        )
+        self._mod = None
+        super(LazyLoader, self).__init__(self._module_name)
+
+    def _load(self):
+        if self._mod is None:
+            self._mod = importlib.import_module(
+                self._module_name
+                )
+        return self._mod
+
+    def __getattr__(self, attrb):
+        try:
+            return getattr(self._load(), attrb)
+        except ModuleNotFoundError:
+            pass
+
+    def __dir__(self):
+        return dir(self._load())
